SPRING CLEANING (lint, refactor,...)
