+++ TicketDesk/ngWebClient-StandAlone/src/app/ticket-action-box/ticket-action-box.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { Ticket } from '../models/data';
+import { Ticket } from '../models/ticket';
 import { TicketActionEnum } from '../models/ticket-actions.constants';
 
 @Component({
