+++ TicketDesk/ngWebClient-StandAlone/src/app/ticket-detail-editor/ticket-detail-editor.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 import { SchemaService } from '../services/schema.service';
 import { SubmitTicketService } from '../services/submit-ticket.service';
-import { Ticket, BLANK_TICKET } from '../models/data';
+import { Ticket, BLANK_TICKET } from '../models/ticket';
 import { AttachFileComponent } from '../attach-file/attach-file.component';
 @Component({
   selector: 'app-ticket-detail-editor',
